# openapi.yaml
openapi: 3.1.0
info:
  description: "This is the REST API documentation of the Spring Petclinic backend.\
    \ If authentication is enabled, use admin/admin when calling the APIs"
  termsOfService: https://github.com/spring-petclinic/petclinic-rest/blob/master/terms.txt
  title: REST Petclinic Backend API
  version: "1.0"
servers:
- description: Generated server url
  url: http://localhost
paths:
...
  /api/specialties:
    get:
      operationId: listSpecialties
      responses:
        "200":
          content:
            application/json:
              examples:
                sample:
                  description: sample
                  value:
                  - id: 1
                    name: radiology
                    description: "x-rays, broken bones, imaging"
                  - id: 2
                    name: surgery
                    description: "operations, spaying, neutering"
                  - id: 3
                    name: dentistry
                    description: "teeth cleaning, extractions"
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpecialtyDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
    post:
      operationId: addSpecialty
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
...
  /api/specialties/{specialtyId}:
    delete:
      operationId: deleteSpecialty
      parameters:
      - in: path
        name: specialtyId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
    get:
      operationId: getSpecialty
      parameters:
      - in: path
        name: specialtyId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpecialtyDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
    put:
      operationId: updateSpecialty
      parameters:
      - in: path
        name: specialtyId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
...
  /api/vets:
    get:
      operationId: listVets
      responses:
        "200":
          content:
            application/json:
              examples:
                sample:
                  description: sample
                  value:
                  - id: 1
                    firstName: James
                    lastName: Carter
                    specialties: []
                  - id: 2
                    firstName: Helen
                    lastName: Leary
                    specialties:
                    - id: 1
                      name: radiology
                      description: x-rays and imaging
                  - id: 3
                    firstName: Linda
                    lastName: Douglas
                    specialties:
                    - id: 2
                      name: surgery
                      description: operations
                    - id: 3
                      name: dentistry
                      description: teeth
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VetDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - vet-rest-controller
    post:
      operationId: addVet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VetDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - vet-rest-controller
  /api/vets/{vetId}:
    delete:
      operationId: deleteVet
      parameters:
      - in: path
        name: vetId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - vet-rest-controller
    get:
      operationId: getVet
      parameters:
      - in: path
        name: vetId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VetDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - vet-rest-controller
    put:
      operationId: updateVet
      parameters:
      - in: path
        name: vetId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VetDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - vet-rest-controller
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
    SpecialtyDto:
      type: object
      properties:
        description:
          type: string
          description: The section that identifies this specialty (symptoms); vectorized
            into the chatbot RAG.
          example: "limping, broken bone, swollen leg, can't bear weight"
          maxLength: 4000
          minLength: 0
        id:
          type: integer
          format: int32
          description: The ID of the specialty.
          example: 1
          minimum: 0
          readOnly: true
        name:
          type: string
          description: The name of the specialty.
          example: radiology
          maxLength: 80
          minLength: 1
      required:
      - id
      - name
...
    VetDto:
      type: object
      properties:
        firstName:
          type: string
          description: The first name of the vet.
          example: James
          maxLength: 30
          minLength: 1
          pattern: \w+
        id:
          type: integer
          format: int32
          description: The ID of the vet.
          example: 1
          minimum: 0
          readOnly: true
        lastName:
          type: string
          description: The last name of the vet.
          example: Carter
          maxLength: 30
          minLength: 1
          pattern: \w+
        specialties:
          type: array
          description: The specialties of the vet.
          items:
            $ref: "#/components/schemas/SpecialtyDto"
      required:
      - firstName
      - id
      - lastName
      - specialties
...